# Parallelize RBI parsing to speed up the `check-shims` command