### Motivation

Running `tapioca check-shims` can take up to 7min in Core (~13 when running on CI). The culprit is parsing the RBI files.

This PR proposes to speed up the process by parallelizing RBI parsing.

### Implementation

I reused `Executor` that we already use to parallelize RBI generation.

### Tests

No behavior change.

I tested the speed-up on Tapioca itself:

Before:

```shell
$ time bundle exec tapioca check-shims

Loading Sorbet payload...  Done (5.36s)
Loading shim RBIs from sorbet/rbi/shims...  Done (0.01s)
Loading gem RBIs from sorbet/rbi/gems...  Done (17.93s)
Looking for duplicates...  Done (0.02s)

real	0m24.291s
user	0m23.251s
sys	0m0.564s
```

After:

```shell
$ time bundle exec tapioca check-shims -w 10
Loading Sorbet payload...  Done (1.51s)
Loading shim RBIs from sorbet/rbi/shims...  Done (0.21s)
Loading gem RBIs from sorbet/rbi/gems...  Done (7.16s)
Looking for duplicates...  Done (0.03s)

real	0m9.561s
user	0m31.153s
sys	0m3.086s
```

So a > 50% improvement here.

I also tested on Core:

Before:

```shell
$ time bundle exec tapioca check-shims
Loading Sorbet payload...  Done (10.05s)
Loading shim RBIs from sorbet/rbi/shims...  Done (1.32s)
Loading gem RBIs from sorbet/rbi/gems...  Done (165.56s)
Loading dsl RBIs from sorbet/rbi/dsl...  Done (227.36s)
Loading annotation RBIs from sorbet/rbi/annotations...  Done (0.08s)
Looking for duplicates...  Done (0.5s)

real	6m52.147s
user	6m48.205s
sys	0m3.855s
```

After:

```shell
$ time bundle exec tapioca check-shims -w 10
Loading Sorbet payload...  Done (6.49s)
Loading shim RBIs from sorbet/rbi/shims...  Done (3.36s)
Loading gem RBIs from sorbet/rbi/gems...  Done (105.36s)
Loading dsl RBIs from sorbet/rbi/dsl...  Done (78.48s)
Loading annotation RBIs from sorbet/rbi/annotations...  Done (3.02s)
Looking for duplicates...  Done (0.44s)

real	3m24.315s
user	9m19.503s
sys	1m1.437s
```

So a > 50% improvement here too.